pyproject.toml
[tool.poetry]
name = "ht-schema"
version = "0.1.0"
description = ""
authors = ["Mark Baggett <markpbaggett@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jsonschema = "^4.23.0"
yamllint = "^1.35.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
